requirements.txt
langchain
langchain-openai
langchain-community
langgraph
chromadb
unstructured[all-docs]
pillow
pytesseract
pdf2image
python-docx
pydantic
pydantic-settings
python-dotenv
fastapi
uvicorn
python-multipart
opencv-python
gradio
streamlit
transformers
torch
torchvision
sentencepiece
PyMuPDF
sentence-transformers
